+++ cmd/psweb/handlers.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -352,25 +351,16 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 			receiverInFeePPM = receiverInFee * 1_000_000 / int64(peer.AsReceiver.SatsIn)
 		}
 
-		ctx := context.Background()
-
 		for _, asset := range []peerswaprpc.AssetType{peerswaprpc.AssetType_BTC, peerswaprpc.AssetType_LBTC} {
 			for _, operation := range []peerswaprpc.OperationType{peerswaprpc.OperationType_SWAP_IN, peerswaprpc.OperationType_SWAP_OUT} {
-				defaultRate, _ := client.GetDefaultPremiumRate(ctx, &peerswaprpc.GetDefaultPremiumRateRequest{
-					Asset:     asset,
-					Operation: operation,
-				})
+				defaultRate, _ := ps.GetDefaultPremiumRate(client, asset, operation)
 				defaultPremium = append(defaultPremium, Premium{
 					Asset:          int32(asset.Number()),
 					Operation:      int32(operation.Number()),
 					PremiumRatePpm: defaultRate.PremiumRatePpm,
 				})
 
-				peerRate, _ := client.GetPremiumRate(ctx, &peerswaprpc.GetPremiumRateRequest{
-					NodeId:    peer.NodeId,
-					Asset:     asset,
-					Operation: operation,
-				})
+				peerRate, _ := ps.GetPremiumRate(client, peer.NodeId, asset, operation)
 				peerPremium = append(peerPremium, Premium{
 					Asset:          int32(asset.Number()),
 					Operation:      int32(operation.Number()),
@@ -1913,11 +1903,9 @@ func submitHandler(w http.ResponseWriter, r *http.Request) {
 		case "setPremium":
 			nextPage := r.FormValue("nextPage")
 
-			premiumRatePpm, err := strconv.ParseInt(r.FormValue("premium"), 10, 64)
-			if err != nil {
-				redirectWithError(w, r, nextPage, err)
-				return
-			}
+			isDeleted := r.FormValue("premium") == ""
+
+			premiumRatePpm, _ := strconv.ParseInt(r.FormValue("premium"), 10, 64)
 
 			asset, err := strconv.ParseInt(r.FormValue("asset"), 10, 32)
 			if err != nil {
@@ -1933,26 +1921,33 @@ func submitHandler(w http.ResponseWriter, r *http.Request) {
 
 			nodeId := r.FormValue("peerNodeId")
 			which := "Default"
+			what := "deleted"
 			var result *peerswaprpc.PremiumRate
 
 			if nodeId == "" {
-				result, err = client.UpdateDefaultPremiumRate(context.Background(), &peerswaprpc.UpdateDefaultPremiumRateRequest{
-					Rate: &peerswaprpc.PremiumRate{
-						Asset:          peerswaprpc.AssetType(asset),
-						Operation:      peerswaprpc.OperationType(operation),
-						PremiumRatePpm: premiumRatePpm,
-					},
+				result, err = ps.UpdateDefaultPremiumRate(client, &peerswaprpc.PremiumRate{
+					Asset:          peerswaprpc.AssetType(asset),
+					Operation:      peerswaprpc.OperationType(operation),
+					PremiumRatePpm: premiumRatePpm,
 				})
+				what = fmt.Sprintf("updated to %d", result.GetPremiumRatePpm())
 			} else {
 				which = "Peer"
-				result, err = client.UpdatePremiumRate(context.Background(), &peerswaprpc.UpdatePremiumRateRequest{
-					NodeId: nodeId,
-					Rate: &peerswaprpc.PremiumRate{
-						Asset:          peerswaprpc.AssetType(asset),
-						Operation:      peerswaprpc.OperationType(operation),
-						PremiumRatePpm: premiumRatePpm,
-					},
-				})
+				if isDeleted {
+					result, err = ps.DeletePremiumRate(client, nodeId,
+						&peerswaprpc.PremiumRate{
+							Asset:     peerswaprpc.AssetType(asset),
+							Operation: peerswaprpc.OperationType(operation),
+						})
+				} else {
+					result, err = ps.UpdatePremiumRate(client, nodeId,
+						&peerswaprpc.PremiumRate{
+							Asset:          peerswaprpc.AssetType(asset),
+							Operation:      peerswaprpc.OperationType(operation),
+							PremiumRatePpm: premiumRatePpm,
+						})
+					what = fmt.Sprintf("updated to %d", result.GetPremiumRatePpm())
+				}
 			}
 
 			if err != nil || result == nil {
@@ -1961,7 +1956,7 @@ func submitHandler(w http.ResponseWriter, r *http.Request) {
 			}
 
 			// all good, display confirmation
-			msg := fmt.Sprintf("%s %s %s premium rate updated to %d", which, result.Asset.String(), result.Operation.String(), result.GetPremiumRatePpm())
+			msg := fmt.Sprintf("%s %s %s premium rate %s", which, result.Asset.String(), result.Operation.String(), what)
 			http.Redirect(w, r, nextPage+"msg="+msg, http.StatusSeeOther)
 			return