+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 1.7.9
 
-- Add premium PPM limit for new swaps
+- Enable setting premium rates and display those of peers
+- Add premium PPM limit for new swaps, blank by default
 - Remove words "Liquid balance" from telegram backups for privacy
+- Reduce L-BTC reserve to 170 sats for max swap calculation
 
 ## 1.7.8